+++ components/script/dom/activation.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use script_layout_interface::message::ReflowGoal;
+use script_layout_interface::ReflowGoal;
 
 use crate::dom::element::Element;
 use crate::dom::event::Event;
